+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntityProjectileBaseComponentImpl.java
@@ -15,7 +15,7 @@ public class EntityProjectileBaseComponentImpl extends EntityBaseComponentImpl {
     }
 
     @Override
-    public AABBdc getAABB() {
+    public AABBdc getBaseAABB() {
         return new AABBd(-0.125, 0.0, -0.125, 0.125, 0.25, 0.125);
     }
 }